// src/shared/enums/actions.enum.ts
export enum ActionsEnums {
	CREATED = "created",
	UPDATED = "updated",
	DELETED = "deleted",
	NOT_FOUND = "not found",
}

export enum ActionsErrorsEnums {
	CREATING = "Error creating",
	UPDATING = "Error updating",
	DELETING = "Error deleting",
	FETCHING = "Error fetching",
	NOT_FOUND = "Not Found",
	INVALID_ID = "Invalid ID",
}